requirements.txt
azure.cosmos
python-dateutil
python-dotenv